requirements/prod.txt
# Frameworks
fastapi==0.103.1

# FastAPI addons
fastapi-users==12.1.2
starlette-csrf==3.0.0
fastapi-users-db-sqlalchemy==6.0.1
starlette-admin==0.11.2

# Validation
pydantic==2.4.0
pydantic-settings==2.0.3

# Redis
aioredis==2.0.1

# Database
alembic==1.12.0
sqlalchemy==2.0.21
fastapi-async-sqlalchemy==0.6.0
psycopg2-binary==2.9.7
redis==5.0.0
asyncpg==0.28.0

# Server
uvicorn==0.23.2

# Other
greenlet==2.0.2
